requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
